+++ fms-cim-server/src/main/java/org/fms/cim/server/webapp/uas/action/PSysNodeAction.java
@@ -2,16 +2,19 @@
  * 系统节点
  * Author :
  * Date :
- * Title : org.fms.eis.webapp.action.PSysNodeAction.java
+ * Title : org.fms.cim.server.webapp.uas.action.PSysNodeAction.java
  **/
 package org.fms.cim.server.webapp.uas.action;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
+import com.fasterxml.jackson.core.JsonParseException;
+import com.fasterxml.jackson.databind.JsonMappingException;
+import com.riozenc.titanTool.common.json.utils.JSONUtil;
+import com.riozenc.titanTool.spring.web.http.HttpResult;
+import com.riozenc.titanTool.spring.web.http.HttpResultPagination;
+import org.apache.commons.collections.IterableMap;
+import org.fms.cim.common.service.IPChnlGpDasRelaService;
 import org.fms.cim.common.service.IPSysNodeService;
-import org.fms.cim.common.vo.uas.PDaserverGroupVO;
-import org.fms.cim.common.vo.uas.PSysNodeVO;
+import org.fms.cim.common.vo.uas.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -20,8 +23,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.riozenc.titanTool.spring.web.http.HttpResult;
-import com.riozenc.titanTool.spring.web.http.HttpResultPagination;
+import java.io.IOException;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 @ControllerAdvice
 @RequestMapping("PSysNode")
@@ -31,6 +36,15 @@ public class PSysNodeAction {
     @Qualifier("PSysNodeServiceImpl")
     private IPSysNodeService pSysNodeService;
 
+    @Autowired
+    @Qualifier("PChnlGpDasRelaServiceImpl")
+    private IPChnlGpDasRelaService pChnlGpDasRelaService;
+
+    /**
+     * 名称和A网地址不允许重复
+     *
+     * @return
+     */
     @ResponseBody
     @PostMapping(params = "method=insert")
     public HttpResult<?> insert(@RequestBody PSysNodeVO pSysNodeVO) {
@@ -50,6 +64,11 @@ public class PSysNodeAction {
         }
     }
 
+    /**
+     * 名称和A网地址不允许重复
+     *
+     * @return
+     */
     @ResponseBody
     @PostMapping(params = "method=update")
     public HttpResult<?> update(@RequestBody PSysNodeVO pSysNodeVO) {
@@ -69,12 +88,33 @@ public class PSysNodeAction {
         }
     }
 
+    /**
+     * 已关联通道组不允许删除
+     *
+     * @return
+     */
     @ResponseBody
     @PostMapping(params = "method=delete")
     public HttpResult<?> delete(@RequestBody List<PSysNodeVO> deleteList) throws Exception {
-        //已关联通道组不允许删除--此处在数据库处理
-        HttpResult httpResult = pSysNodeService.deleteList(deleteList);
-        return httpResult;
+        if (deleteList != null) {
+            if (deleteList.size() > 0) {
+                String value = "";
+                for (PSysNodeVO item : deleteList) {
+                    value += item.getId() + ",";
+                }
+                value = value.substring(0, value.length() - 1);
+                List<PChnlGpDasRelaVO> relList = pChnlGpDasRelaService.findByRelSysNode(value);
+                if (relList.size() > 0) {
+                    return new HttpResult<String>(HttpResult.ERROR, "已关联通道组不允许删除", null);
+                } else {
+                    return pSysNodeService.deleteList(deleteList);
+                }
+            } else {
+                return new HttpResult<String>(HttpResult.ERROR, "暂无要删除的内容", null);
+            }
+        } else {
+            return new HttpResult<String>(HttpResult.ERROR, "参数传递错误", null);
+        }
     }
 
     @ResponseBody
@@ -123,6 +163,7 @@ public class PSysNodeAction {
 
     /**
      * 批量更新主机的采集机组
+     *
      * @param pDaserverGroupVO 采集机组以及其选中的子集主机
      * @return
      */
@@ -133,7 +174,8 @@ public class PSysNodeAction {
             PSysNodeVO pSysNodeVO = new PSysNodeVO();
             pSysNodeVO.setDaGroup(pDaserverGroupVO.getId());
             List<PSysNodeVO> sysNodeVOList = pSysNodeService.findByWhere(pSysNodeVO);//更新前
-            List<Long> childIDList = pDaserverGroupVO.getListSysNodeVO().stream().map(PSysNodeVO::getId).collect(Collectors.toList());//设置选中的
+            List<Long> childIDList = pDaserverGroupVO.getListSysNodeVO()
+                    .stream().map(PSysNodeVO::getId).collect(Collectors.toList());//设置选中的
             if (sysNodeVOList != null && sysNodeVOList.size() > 0) {
                 for (PSysNodeVO item : sysNodeVOList) {
                     Long daserverGroupID = childIDList.contains(item.getId()) ? item.getDaGroup() : null;
